fix(tests/unit): fix pytest warning
```
=============================== warnings summary ===============================
build/tox/py313-check/lib/python3.13/site-packages/pygments/formatters/other.py:129: PytestCollectionWarning: cannot collect test class 'TestcaseFormatter' because it has a __init__ constructor (from: tests/unit/strictdoc/export/rst/test_strictdoc_lexer.py)
    class TestcaseFormatter(Formatter):
```
